+++ src/app/layout.tsx
@@ -9,7 +9,7 @@ import {
   GoToBottomButton,
 } from "@/app/_common";
 import Providers from "./providers";
-import LocalStorageSetter from "./local-storage-setter";
+import StorageSetter from "./storage-setter";
 import "./globals.css";
 
 /** Next.js에서 제공하는 localFont 기능으로, 폰트는 Pretendard 사용 */
@@ -37,7 +37,7 @@ export default function RootLayout({
         <Analytics />
         <SpeedInsights />
         <Providers>
-          <LocalStorageSetter />
+          <StorageSetter />
           <div className="relative w-full h-full min-h-screen bg-dark">
             <Header />
             <div className="flex flex-col w-full max-w-3xl h-full min-h-screen bg-dark py-12 mx-auto">
